Complete code refactoring re specificities.
Specificities are completely parameterised and for the most part are no longer hardcoded. though annoyingly the slightly different behaviour that books have relative to the other parts makes referencing specificities required, at least for the time being.

Additionally added a new algorithm for checking when a new page is detected.